[MTIA] [PyTorch] [Export] Allow setting names for lifted constants before tracing. (#162926)
Summary:

The lift_constants_pass will set the name of lifted to constants to module_path_{lifted_tensor_counter}. This makes it very difficult for a compiler or runtime to map certain constants to what the placeholder value is in the IR. 

This PR adds a new API to Torch Export allowing users to set the names of certain tensors. The lift_constants_pass then respects this name. 

This is not supported for strict=True or for naming FakeTensors.

```
        tensor = torch.tensor([1, 2, 3])
        result = torch.export._name_tensor_constants.set_tensor_name(tensor, "my_custom_tensor")
        self.assertIs(result, tensor)  # Should return the same tensor
```

Before:
```
 %model_layers_0_self_attn_attn_lifted_tensor_0 : bf16[2, 346732, 16, 8, 64][num_users=2] = placeholder[target='model_layers_0_self_attn_attn_lifted_tensor_0']
```

```
 kv_cache = torch.export._name_tensor_constants.set_tensor_name(kv_cache, "kv_cache")
```

After:


```
%model_layers_0_self_attn_attn_kv_cache : bf16[2, 346544, 16, 8, 64][num_users=2] = placeholder[target='model_layers_0_self_attn_attn_kv_cache']
```

Test Plan:
Added unit tests

Rollback Plan:

Differential Revision: D82387095